requirement.txt
asgiref==3.10.0
boto3==1.40.55
botocore==1.40.55
Django==4.2.25
django-storages==1.14.6
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jmespath==1.0.1
pillow==11.3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==1.26.20
